Add remove rounded button in helper plugin
