+++ packages/core/helper-plugin/lib/src/icons/RemoveRoundedButton/RemoveRoundedButton.stories.mdx
@@ -1,0 +1,18 @@
+<!--- RemoveRoundedButton.stories.mdx --->
+
+import { Meta, Canvas, Story } from '@storybook/addon-docs';
+import RemoveRoundedButton from './index';
+
+<Meta title="buttons/RemoveRoundedButton" />
+
+# RemoveRoundedButton
+
+This component is used in order to remove a row.
+
+## Usage
+
+<Canvas>
+  <Story name="base">
+    <RemoveRoundedButton label="Remove me" />
+  </Story>
+</Canvas>
